requirements.txt
bluepy==1.3.0
PyYAML==5.4.1
